Update endpoint for doirequest
